PRSD-1392: Combine PropertyComplianceUpdate single page test files
